soliqai/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
asyncpg
python-multipart
pymupdf
pytesseract
chromadb
ollama
openai
python-dotenv
pydantic-settings
bcrypt
pyjwt
python-jose[cryptography]
passlib[bcrypt]
pdf2image
python-docx
